Accounting team requirement - Landlord Payment

+++ resources/views/landlord_payments/single_table.blade.php
@@ -51,7 +51,9 @@
                         <td>
                             <a class="btn btn-success" href="{{ route( Str::camel($layer) . '.show', $object['id']) }}">查看</a>
                             <a class="btn btn-primary" href="{{ route( Str::camel($layer) . '.edit', $object['id']) }}">編輯</a>
-                            <a class="btn btn-danger jquery-postback" data-method="delete" href="{{ route( Str::camel($layer) . '.show', $object['id']) }}">刪除</a>
+                            @if (! ($layer === 'buildings' && request()->route()->getName() === 'landlordPayments.show'))
+                                <a class="btn btn-danger jquery-postback" data-method="delete" href="{{ route( Str::camel($layer) . '.show', $object['id']) }}">刪除</a>
+                            @endif
                         </td>
                     </tr>
                 </tbody>
